Add nocompatible for vi
This command makes it so VIM does not need to keep compatability with
old school Vi. I'm sure there might be a use for this for some people
but since this a new build, and most addon/features need this set.
You can use `:help 'compatible'` to read more about this. Also note
that nocompatible is automatically turned on for us but this is used
to make sure it is always on.