Add line numbers to ruler and sidebar.
Vim now has sidebar numbers that are realtive to the cursor. You
can also see in the ruler the line number. The first number is the
line your are on and the second number after the comma is the
number where the cursor is. I am adding all three settings to one
commit as they are related.