Add Base16 ocean color scheme
This was very difficult to get set up and even now that it's done
I don't think it's completely there. This set up required me to make
the switch to iTerm instead Terminal. There are conflicts with 256 and
base16 themes with Terminal and it was simpler to switch applications.

I created a colors dir and included the `.vim` files needed to support
the ocean theme(as well as some others). I also included the
base16-shell scripts in my `~/.config` dir where I could test that
my 256 colors were indeed working. In iTerm I double clicked on
downloaded `.itermcolor` files to include the needed ocean color
themes for iTerm and created a new template.

This has been the most frustrating thing thus far and I'm sure this
won't be the last time that color themes will create issues for me.