Set up Pathogen and first vim package
I did this because there are vim packages that will make my life a bit more
fun when it comes to vim. I don't want it to be another text editor but I do
like some of the features that can be found on text editors.

I installed my first git submodule for pathogen. It makes a type of sym link
to the repository of the project. Now I can just run `git submodule init` and
`git submodule update` to get all the settings for vim and its packages onto
my machine.