change file and text search
I did this because I want to use file name searching and text searching. I
think that using f and F for my file searching was not what I liked so I
switched it to P so I remeber its ctrl p vim that's doing the work.

Also added a map for ack so I don't have type ack evertime.