Add open alternate and correct ack
I did this because I want to be able to switch between my tests and class
quickly through a keyboard shortcut.

I also wanted to have ack use ack! so that it would stop opening files
automatically when it finds a file. I want to initiate the opening in my
buffer.