Add line memory for vim and update test recall
I did this because switching between code and my tests were starting me at the
top of my file all the time. This was annoying and the internet and people
have this in their vimrc for remembering the line that the cursor is on.

Test recall will use `bundle exec rspec` when I run my tests. This is helpful
for running the tests in my rails app.