Switch binding and remove cucumber tests running for now
I did this because I am not using cucumber currently and it's getting in the
way of my clear search highlight mapping. I will decide what I want the
binding to be when I need to use cucumber.

I am switching to lowercase `t` for running nearest test because I use that
more than the whole file. I'm lazy and don't want to press shift so much or
accidentally press lowercase when I mean to do capital.