Add command history and show hidden in ctrlp
I did this because I want to see some config dot files that exist in my
projects when I search for them with ctrl p. also when I am using the ex
command history I want to not have to hit arrows and ctrl p and n to move up
and down is a common vim thing.